backend_server/requirements.txt
jsonrpclib-pelix
pymongo
pika